Using only .mts tests, ignore generated .mjs
