+++ feature_tests/js/test/.gitignore
@@ -1,0 +1,1 @@
+*.mjs
